move module cards to PipettesAndModules component
