+++ app/src/pages/Devices/DeviceDetails/__tests__/DeviceDetails.test.tsx
@@ -4,21 +4,18 @@ import { renderWithProviders } from '@opentrons/components'
 
 import { i18n } from '../../../../i18n'
 import { mockConnectableRobot } from '../../../../redux/discovery/__fixtures__'
-import {
-  mockFetchModulesSuccessActionPayloadModules,
-  mockMagneticModule,
-} from '../../../../redux/modules/__fixtures__'
+import { mockFetchModulesSuccessActionPayloadModules } from '../../../../redux/modules/__fixtures__'
 import {
   useAttachedModules,
   useRobot,
 } from '../../../../organisms/Devices/hooks'
+import { PipettesAndModules } from '../../../../organisms/Devices/PipettesAndModules'
 import { RobotOverview } from '../../../../organisms/Devices/RobotOverview'
 import { DeviceDetails } from '..'
-import { ModuleCard } from '../../../../organisms/Devices/ModuleCard'
 
 jest.mock('../../../../organisms/Devices/hooks')
+jest.mock('../../../../organisms/Devices/PipettesAndModules')
 jest.mock('../../../../organisms/Devices/RobotOverview')
-jest.mock('../../../../organisms/Devices/ModuleCard')
 
 const mockUseAttachedModules = useAttachedModules as jest.MockedFunction<
   typeof useAttachedModules
@@ -27,7 +24,9 @@ const mockUseRobot = useRobot as jest.MockedFunction<typeof useRobot>
 const mockRobotOverview = RobotOverview as jest.MockedFunction<
   typeof RobotOverview
 >
-const mockModuleCard = ModuleCard as jest.MockedFunction<typeof ModuleCard>
+const mockPipettesAndModules = PipettesAndModules as jest.MockedFunction<
+  typeof PipettesAndModules
+>
 
 const render = (path = '/') => {
   return renderWithProviders(
@@ -47,7 +46,7 @@ describe('DeviceDetails', () => {
     )
     mockUseRobot.mockReturnValue(null)
     mockRobotOverview.mockReturnValue(<div>Mock RobotOverview</div>)
-    mockModuleCard.mockReturnValue(<div>Mock ModuleCard</div>)
+    mockPipettesAndModules.mockReturnValue(<div>Mock PipettesAndModules</div>)
   })
   afterEach(() => {
     jest.resetAllMocks()
@@ -65,11 +64,11 @@ describe('DeviceDetails', () => {
 
     getByText('Mock RobotOverview')
   })
-  it('renders a Module card', () => {
+
+  it('renders PipettesAndModules when a robot is found', () => {
     mockUseRobot.mockReturnValue(mockConnectableRobot)
-    mockUseAttachedModules.mockReturnValue([mockMagneticModule])
     const [{ getByText }] = render('/devices/otie')
 
-    getByText('Mock ModuleCard')
+    getByText('Mock PipettesAndModules')
   })
 })
